requirements.txt
flask
transformers
torch
requests
openai
python-dotenv
accelerate
sentencepiece
bitsandbytes
huggingface_hub
protobuf
traceback